DOCS:문서 수정 및 삭제
