+++ .env
@@ -1,5 +1,4 @@
-REACT_APP_KAKAO_REST_KEY=84d4cb0f0b38976fab6edba825421247
-
+REACT_APP_KAKAO_REST_KEY=3b06fd5ed713d988623e3713a5264e3e
 REACT_APP_KAKAO_REDIRECT_URI=http://localhost:3000/api/user/kakao/callback
 REACT_APP_SEARCH_PLANT=/api/plants
 REACT_APP_MY_PLANT=/api/beginner/plant/my
